Moved tests to subnamespaces to match tested classes
