+++ test/Config/Option/Database/DatabaseUserConfigOptionTest.php
@@ -2,29 +2,30 @@
 
 declare(strict_types=1);
 
-namespace ShlinkioTest\Shlink\Installer\Config\Option;
+namespace ShlinkioTest\Shlink\Installer\Config\Option\Database;
 
 use PHPUnit\Framework\TestCase;
+use Prophecy\Argument;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Shlinkio\Shlink\Config\Collection\PathCollection;
-use Shlinkio\Shlink\Installer\Config\Option\Database\DatabaseNameConfigOption;
+use Shlinkio\Shlink\Installer\Config\Option\Database\DatabaseUserConfigOption;
 use Symfony\Component\Console\Style\StyleInterface;
 
-class DatabaseNameConfigOptionTest extends TestCase
+class DatabaseUserConfigOptionTest extends TestCase
 {
     use ProphecyTrait;
 
-    private DatabaseNameConfigOption $configOption;
+    private DatabaseUserConfigOption $configOption;
 
     public function setUp(): void
     {
-        $this->configOption = new DatabaseNameConfigOption();
+        $this->configOption = new DatabaseUserConfigOption();
     }
 
     /** @test */
     public function returnsExpectedConfig(): void
     {
-        $this->assertEquals(['entity_manager', 'connection', 'dbname'], $this->configOption->getConfigPath());
+        $this->assertEquals(['entity_manager', 'connection', 'user'], $this->configOption->getConfigPath());
     }
 
     /** @test */
@@ -32,7 +33,7 @@ class DatabaseNameConfigOptionTest extends TestCase
     {
         $expectedAnswer = 'the_answer';
         $io = $this->prophesize(StyleInterface::class);
-        $ask = $io->ask('Database name', 'shlink')->willReturn($expectedAnswer);
+        $ask = $io->ask('Database username', null, Argument::any())->willReturn($expectedAnswer);
 
         $answer = $this->configOption->ask($io->reveal(), new PathCollection());
