+++ test/Config/Option/Redirect/BaseUrlRedirectConfigOptionTest.php
@@ -2,30 +2,30 @@
 
 declare(strict_types=1);
 
-namespace ShlinkioTest\Shlink\Installer\Config\Option;
+namespace ShlinkioTest\Shlink\Installer\Config\Option\Redirect;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Shlinkio\Shlink\Config\Collection\PathCollection;
-use Shlinkio\Shlink\Installer\Config\Option\UrlShortener\ShortDomainHostConfigOption;
+use Shlinkio\Shlink\Installer\Config\Option\Redirect\BaseUrlRedirectConfigOption;
 use Symfony\Component\Console\Style\StyleInterface;
 
-class ShortDomainHostConfigOptionTest extends TestCase
+class BaseUrlRedirectConfigOptionTest extends TestCase
 {
     use ProphecyTrait;
 
-    private ShortDomainHostConfigOption $configOption;
+    private BaseUrlRedirectConfigOption $configOption;
 
     public function setUp(): void
     {
-        $this->configOption = new ShortDomainHostConfigOption();
+        $this->configOption = new BaseUrlRedirectConfigOption();
     }
 
     /** @test */
     public function returnsExpectedConfig(): void
     {
-        $this->assertEquals(['url_shortener', 'domain', 'hostname'], $this->configOption->getConfigPath());
+        $this->assertEquals(['not_found_redirects', 'base_url'], $this->configOption->getConfigPath());
     }
 
     /** @test */
@@ -33,7 +33,12 @@ class ShortDomainHostConfigOptionTest extends TestCase
     {
         $expectedAnswer = 'the_answer';
         $io = $this->prophesize(StyleInterface::class);
-        $ask = $io->ask('Default domain for generated short URLs', null, Argument::any())->willReturn($expectedAnswer);
+        $ask = $io->ask(
+            'Custom URL to redirect to when a user hits Shlink\'s base URL (If no value is provided, the '
+            . 'user will see a default "404 not found" page)',
+            null,
+            Argument::any(),
+        )->willReturn($expectedAnswer);
 
         $answer = $this->configOption->ask($io->reveal(), new PathCollection());
