+++ test/Config/Option/Redirect/InvalidShortUrlRedirectConfigOptionTest.php
@@ -2,30 +2,30 @@
 
 declare(strict_types=1);
 
-namespace ShlinkioTest\Shlink\Installer\Config\Option;
+namespace ShlinkioTest\Shlink\Installer\Config\Option\Redirect;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Shlinkio\Shlink\Config\Collection\PathCollection;
-use Shlinkio\Shlink\Installer\Config\Option\Redirect\BaseUrlRedirectConfigOption;
+use Shlinkio\Shlink\Installer\Config\Option\Redirect\InvalidShortUrlRedirectConfigOption;
 use Symfony\Component\Console\Style\StyleInterface;
 
-class BaseUrlRedirectConfigOptionTest extends TestCase
+class InvalidShortUrlRedirectConfigOptionTest extends TestCase
 {
     use ProphecyTrait;
 
-    private BaseUrlRedirectConfigOption $configOption;
+    private InvalidShortUrlRedirectConfigOption $configOption;
 
     public function setUp(): void
     {
-        $this->configOption = new BaseUrlRedirectConfigOption();
+        $this->configOption = new InvalidShortUrlRedirectConfigOption();
     }
 
     /** @test */
     public function returnsExpectedConfig(): void
     {
-        $this->assertEquals(['not_found_redirects', 'base_url'], $this->configOption->getConfigPath());
+        $this->assertEquals(['not_found_redirects', 'invalid_short_url'], $this->configOption->getConfigPath());
     }
 
     /** @test */
@@ -34,7 +34,7 @@ class BaseUrlRedirectConfigOptionTest extends TestCase
         $expectedAnswer = 'the_answer';
         $io = $this->prophesize(StyleInterface::class);
         $ask = $io->ask(
-            'Custom URL to redirect to when a user hits Shlink\'s base URL (If no value is provided, the '
+            'Custom URL to redirect to when a user hits an invalid short URL (If no value is provided, the '
             . 'user will see a default "404 not found" page)',
             null,
             Argument::any(),
