+++ test/Config/Option/UrlShortener/ShortCodeLengthOptionTest.php
@@ -2,41 +2,41 @@
 
 declare(strict_types=1);
 
-namespace ShlinkioTest\Shlink\Installer\Config\Option;
+namespace ShlinkioTest\Shlink\Installer\Config\Option\UrlShortener;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Shlinkio\Shlink\Config\Collection\PathCollection;
-use Shlinkio\Shlink\Installer\Config\Option\Redirect\InvalidShortUrlRedirectConfigOption;
+use Shlinkio\Shlink\Installer\Config\Option\UrlShortener\ShortCodeLengthOption;
 use Symfony\Component\Console\Style\StyleInterface;
 
-class InvalidShortUrlRedirectConfigOptionTest extends TestCase
+class ShortCodeLengthOptionTest extends TestCase
 {
     use ProphecyTrait;
 
-    private InvalidShortUrlRedirectConfigOption $configOption;
+    private ShortCodeLengthOption $configOption;
 
     public function setUp(): void
     {
-        $this->configOption = new InvalidShortUrlRedirectConfigOption();
+        $this->configOption = new ShortCodeLengthOption();
     }
 
     /** @test */
     public function returnsExpectedConfig(): void
     {
-        $this->assertEquals(['not_found_redirects', 'invalid_short_url'], $this->configOption->getConfigPath());
+        $this->assertEquals(['url_shortener', 'default_short_codes_length'], $this->configOption->getConfigPath());
     }
 
     /** @test */
     public function expectedQuestionIsAsked(): void
     {
-        $expectedAnswer = 'the_answer';
+        $expectedAnswer = 5;
         $io = $this->prophesize(StyleInterface::class);
         $ask = $io->ask(
-            'Custom URL to redirect to when a user hits an invalid short URL (If no value is provided, the '
-            . 'user will see a default "404 not found" page)',
-            null,
+            'What is the default length you want generated short codes to have? (You will still be able to override '
+            . 'this on every created short URL)',
+            '5',
             Argument::any(),
         )->willReturn($expectedAnswer);
